api-cotizaciones/requirements.txt
fastapi
uvicorn
pydantic
reportlab
requests
python-docx